try to use a global FLOAT

+++ src/hydro/eos/cubic_spline.rs
@@ -1,37 +1,39 @@
 use boxarray::boxarray;
 
-pub fn cubic(x: f64, p: &[f64; 4]) -> f64 {
+use crate::FLOAT;
+
+pub fn cubic(x: FLOAT, p: &[FLOAT; 4]) -> FLOAT {
     p[0] + x * (p[1] + x * (p[2] + x * p[3]))
 }
 
-pub fn cubic_diff(x: f64, p: &[f64; 4]) -> f64 {
+pub fn cubic_diff(x: FLOAT, p: &[FLOAT; 4]) -> FLOAT {
     p[1] + x * (2.0 * p[2] + x * 3.0 * p[3])
 }
 
-pub fn cubic_diff2(x: f64, p: &[f64; 4]) -> f64 {
+pub fn cubic_diff2(x: FLOAT, p: &[FLOAT; 4]) -> FLOAT {
     2.0 * p[2] + x * 6.0 * p[3]
 }
 
-pub fn cubic_diff3(_x: f64, p: &[f64; 4]) -> f64 {
+pub fn cubic_diff3(_x: FLOAT, p: &[FLOAT; 4]) -> FLOAT {
     6.0 * p[3]
 }
 
-pub fn cubic_log(x: f64, p: &[f64; 4]) -> f64 {
+pub fn cubic_log(x: FLOAT, p: &[FLOAT; 4]) -> FLOAT {
     cubic(x, p).exp()
 }
 
 pub fn cubic_spline_0<const N: usize, const M: usize>(
     idx: usize,
     id: usize,
-    arr: &[[f64; M]; N],
-) -> Box<[[f64; 4]; N]> {
+    arr: &[[FLOAT; M]; N],
+) -> Box<[[FLOAT; 4]; N]> {
     let x0 = arr[0][idx];
     let x1 = arr[1][idx];
     let x1x0 = x1 - x0;
     let a = x0 / x1x0;
     let aa = a * a;
 
-    let mut tmp = vec![[0.0f64; 3]; N + 1];
+    let mut tmp = vec![[0.0 as FLOAT; 3]; N + 1];
     tmp[0][0] = 0.0;
     tmp[0][2] = 2.0 * a;
     for i in 0..N {
@@ -72,7 +74,7 @@ pub fn cubic_spline_0<const N: usize, const M: usize>(
     tmp[0][1] -= tmp[1][1] * m01;
     tmp[0][1] /= tmp[0][2];
 
-    let mut pols: Box<[[f64; 4]; N]> = boxarray(0.0);
+    let mut pols: Box<[[FLOAT; 4]; N]> = boxarray(0.0);
     pols[0][0] = tmp[0][0];
     pols[0][1] = a * tmp[0][1];
     pols[0][2] = 3.0 * (tmp[1][0] - tmp[0][0]) - a * (2.0 * tmp[0][1] + tmp[1][1]);
@@ -94,9 +96,9 @@ pub fn cubic_spline_0<const N: usize, const M: usize>(
 
 pub fn cubic_spline<const N: usize, const M: usize>(
     id: usize,
-    arr: &[[f64; M]; N],
-) -> Box<[[f64; 4]; N]> {
-    let mut tmp: Box<[[f64; 4]; N]> = boxarray(0.0);
+    arr: &[[FLOAT; M]; N],
+) -> Box<[[FLOAT; 4]; N]> {
+    let mut tmp: Box<[[FLOAT; 4]; N]> = boxarray(0.0);
     for i in 0..N {
         tmp[i][0] = arr[i][id];
         tmp[i][2] = 4.0;
@@ -121,7 +123,7 @@ pub fn cubic_spline<const N: usize, const M: usize>(
     }
     tmp[0][1] /= tmp[0][2];
 
-    let mut pols: Box<[[f64; 4]; N]> = boxarray(0.0);
+    let mut pols: Box<[[FLOAT; 4]; N]> = boxarray(0.0);
     for i in 0..N - 1 {
         pols[i][0] = tmp[i][0];
         pols[i][1] = tmp[i][1];
